+++ tests/testthat/test-use_dockerfile.R
@@ -3,8 +3,11 @@ test_that("use_dockerfile works", {
   path <- use_dockerfile(save_dir=tempdir())
   testthat::expect_true(file.exists(path))
   
-  path <- use_dockerfile(save_dir=tempdir(), 
-                         force_new = FALSE,
-                         show = TRUE)
-  testthat::expect_true(file.exists(path))
+  out <- testthat::capture_output_lines({
+    path2 <- use_dockerfile(save_dir=tempdir(), 
+                            force_new = FALSE,
+                            show = TRUE)
+  })
+  testthat::expect_true(file.exists(path2))
+  testthat::expect_equal("# ----- R Package Dockerfile -----",out[[1]])
 })